Create dataMigration.sql
This is to resolve an issue I recently faced.

This issue has genesis in fact that when we upgraded our hibernate version from 3.3.2GA  to 4.3. But I guess this issue shall come for upgrade to 3.4 version onwards, But I have not tested that.

This issue is because of @Lob annotation of JPA that we use on our code. We use jboss with hibernate as provider for our JPA. And we use postgres as our database.

Before upgrade, the entities having @Lob annotations, are having data stored in the columns directly. But after upgrade  of hibernate Application, @Lob fields in postgres are stored as OIDs (Only a reference to the data is stored in the column. Actual data is stored by postgres at the referenced location).

Because of this, When I try to work with my application and whenever my application tries to retrieve data from one the tables which had data in text columns before upgrade (which are the results of @Lob annotation)  I get an exception like this:

33m16:24:56,742 WARN  [org.hibernate.engine.jdbc.spi.SqlExceptionHelper] (default task-7) SQL Error: 0, SQLState: 22003
31m16:24:56,742 ERROR [org.hibernate.engine.jdbc.spi.SqlExceptionHelper] (default task-7) Bad value for type long :
.
.
Caused by: javax.persistence.PersistenceException: org.hibernate.exception.DataException: could not execute query 
.
.
To resolve this I had to use the above script.


Some good links to read:

http://www.experts-exchange.com/Software/Server_Software/Application_Servers/Java/Q_24028302.html

http://www.shredzone.de/cilla/page/299/string-lobs-on-postgresql-with-hibernate-36.html#comments

http://stackoverflow.com/questions/14512771/how-do-i-migrate-postgresql-strings-to-a-text-type